Rename `workouts-table-dialog-content-types.ts` file to `types.ts` and update imports
