+++ src/components/workouts-table/workouts-table-dialog/workouts-table-dialog-content/types.ts
